--- front-end/src/global/addressPrefix.ts
@@ -1,10 +1,0 @@
-// Copyright 2019-2020 @Premiurly/polkassembly authors & contributors
-// This software may be modified and distributed under the terms
-// of the Apache-2.0 license. See the LICENSE file for details.
-
-export const addressPrefix: Record<string, number> = {
-	'kusama': 2,
-	'moonbeam': 1284,
-	'moonriver': 1285,
-	'polkadot': 0
-};
